fix(formbuilder, formbuilderhelpers): move helper functions to helper file
